Converting to go now.
Got libgit loading (pain in the ass)

- Need to use the most up-to-date libgit2,
which is not the one installed by homebrew.

Will start working on directory structure again.
May store metadata in a file that points to the
to make the lookups faster.

This file should be artificially inflated
similar to the way that the old files were done
That way having more passwords won't be obvious
by the size of the file, or the number of pw files